Only replace openapi spec on diff

+++ collect-server.azure.sh
@@ -605,9 +605,12 @@ if [[ -f ${indir}/${build_id}/openapi.json ]]; then
         link_name="jellyfin-openapi-unstable.json"
     fi
     mkdir -p ${api_dir}
-    mv ${indir}/${build_id}/openapi.json ${api_dir}/jellyfin-openapi-${api_version}.json
-    rm -f ${api_root}/${link_name}
-    ln -s ${api_dir}/jellyfin-openapi-${api_version}.json ${api_root}/${link_name}
+    if ! diff -q ${indir}/${build_id}/openapi.json ${api_root}/${link_name} &>/dev/null; then
+        # Only replace the OpenAPI spec if they differ
+        mv ${indir}/${build_id}/openapi.json ${api_dir}/jellyfin-openapi-${api_version}.json
+        rm -f ${api_root}/${link_name}
+        ln -s ${api_dir}/jellyfin-openapi-${api_version}.json ${api_root}/${link_name}
+    fi
 fi
 
 # Cleanup
